Jk/cumulus 3204 (#3356)
* Reorder/updated CHANGELOG

* Minor P3 release edits

* Remove unneded test file (duplicate)

* Remove duplicated code

* Remove unused fetchAllRules method

* merge /lib/test-sns-rules.js into lib/rules/test-rulesHelpers.js

* Merge test-createRuleResources into test-rulesHelpers

* Remove update resources tests, add to ruleshelpers test.    Add
missing tests

* Update CHANGELOG

* Remove duplicated test

* Fix lint/formatting concern

* Add updated test scenario

* Fix lint failures
