+++ packages/api/endpoints/rules.js
@@ -18,7 +18,6 @@ const { indexRule, deleteRule } = require('@cumulus/es-client/indexer');
 const { isBadRequestError } = require('../lib/errors');
 const {
   createRuleTrigger,
-  deleteOldEventSourceMappings,
   deleteRuleResources,
   invokeRerun,
   updateRuleTrigger,
@@ -168,9 +167,9 @@ async function put(req, res) {
       translatedRule = await translatePostgresRuleToApiRule(pgRule, knex);
       await indexRule(esClient, translatedRule, process.env.ES_INDEX);
     });
-
-    await deleteOldEventSourceMappings(knex, oldApiRule);
-
+    if (['kinesis', 'sns'].includes(oldApiRule.rule.type)) {
+      await deleteRuleResources(knex, oldApiRule);
+    }
     return res.send(translatedRule);
   } catch (error) {
     log.error('Unexpected error when updating rule:', error);
