+++ packages/api/tests/lib/rules/test-rulesHelpers.js
@@ -2,33 +2,46 @@
 
 const test = require('ava');
 const sinon = require('sinon');
+const omit = require('lodash/omit');
+
 const proxyquire = require('proxyquire');
 const fs = require('fs-extra');
 
 const awsServices = require('@cumulus/aws-client/services');
+const workflows = require('@cumulus/common/workflows');
+
+const SQS = require('@cumulus/aws-client/SQS');
 const { recursivelyDeleteS3Bucket } = require('@cumulus/aws-client/S3');
 const { sqsQueueExists } = require('@cumulus/aws-client/SQS');
 const { randomId, randomString } = require('@cumulus/common/test-utils');
 const {
+  CollectionPgModel,
   destroyLocalTestDb,
+  fakeCollectionRecordFactory,
+  fakeProviderRecordFactory,
   generateLocalTestDb,
   localStackConnectionEnv,
   migrationDir,
+  ProviderPgModel,
   RulePgModel,
+  translateApiRuleToPostgresRule,
   translateApiRuleToPostgresRuleRaw,
 } = require('@cumulus/db');
 
 const { createSqsQueues, fakeRuleFactoryV2 } = require('../../../lib/testUtils');
 const {
+  buildPayload,
+  deleteKinesisEventSources,
   checkForSnsSubscriptions,
   createRuleTrigger,
   deleteRuleResources,
-  deleteOldEventSourceMappings,
+  updateRuleTrigger,
 } = require('../../../lib/rulesHelpers');
 const { getSnsTriggerPermissionId } = require('../../../lib/snsRuleHelpers');
 
 const listRulesStub = sinon.stub();
 
+// TODO remove proxyquire/don't use rulesHelpers require
 const rulesHelpers = proxyquire('../../../lib/rulesHelpers', {
   '@cumulus/api-client/rules': {
     listRules: listRulesStub,
@@ -37,6 +50,8 @@ const rulesHelpers = proxyquire('../../../lib/rulesHelpers', {
     handleScheduleEvent: (payload) => payload,
   },
 });
+
+const { ResourceNotFoundError, resourceNotFoundInfo } = require('../../../lib/errors');
 
 const testDbName = randomString(12);
 
@@ -63,6 +78,21 @@ const getKinesisEventMappings = async () => {
     return awsServices.lambda().listEventSourceMappings(mappingParms).promise();
   });
   return await Promise.all(mappingPromises);
+};
+
+const deleteKinesisEventSourceMappings = async () => {
+  const eventMappings = await getKinesisEventMappings();
+
+  if (!eventMappings) {
+    return Promise.resolve();
+  }
+
+  const allEventMappings = eventMappings[0].EventSourceMappings.concat(
+    eventMappings[1].EventSourceMappings
+  );
+
+  return await Promise.all(allEventMappings.map((e) =>
+    awsServices.lambda().deleteEventSourceMapping({ UUID: e.UUID }).promise()));
 };
 
 test.before(async (t) => {
@@ -110,10 +140,21 @@ test.before(async (t) => {
   t.context.testKnex = knex;
   t.context.testKnexAdmin = knexAdmin;
   t.context.rulePgModel = new RulePgModel();
+  t.context.providerPgModel = new ProviderPgModel();
+  t.context.collectionPgModel = new CollectionPgModel();
 });
 
-test.afterEach.always(() => {
+test.beforeEach(async (t) => {
+  t.context.sandbox = sinon.createSandbox();
+  const topic = await awsServices.sns().createTopic({ Name: randomId('sns') }).promise();
+  t.context.snsTopicArn = topic.TopicArn;
+  await deleteKinesisEventSourceMappings();
+});
+
+test.afterEach.always(async (t) => {
   listRulesStub.reset();
+  t.context.sandbox.restore();
+  await awsServices.sns().deleteTopic({ TopicArn: t.context.snsTopicArn }).promise();
 });
 
 test.after.always(async (t) => {
@@ -511,6 +552,69 @@ test.serial('deleteKinesisEventSources deletes all kinesis event sources', async
   t.is(deletedLogEventMappings.length, 0);
 });
 
+test.serial('deleteKinesisEventSources throws when deleteKinesisEventSource throws', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const {
+    testKnex,
+  } = t.context;
+
+  const params = {
+    rule: {
+      arn: kinesisArn,
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+    state: 'ENABLED',
+    provider: null,
+    collection: null,
+  };
+  const kinesisRule = fakeRuleFactoryV2(params);
+  await t.throwsAsync(
+    deleteKinesisEventSources(
+      testKnex,
+      kinesisRule,
+      {
+        deleteKinesisEventSourceMethod: () => {
+          throw new Error('Test Error');
+        },
+      },
+      { message: 'Test Error' }
+    )
+  );
+});
+
+test.serial('deleteKinesisEventSources does not throw when deleteKinesisEventSource throws a ResourceNotFoundException', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const {
+    testKnex,
+  } = t.context;
+
+  const params = {
+    rule: {
+      arn: kinesisArn,
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+    state: 'ENABLED',
+    provider: null,
+    collection: null,
+  };
+  const kinesisRule = fakeRuleFactoryV2(params);
+  class ResourceNotFoundException extends Error {
+    constructor(...args) {
+      super(...args);
+      this.code = 'ResourceNotFoundException';
+    }
+  }
+  const deleteResult = await deleteKinesisEventSources(testKnex, kinesisRule, {
+    // eslint-disable-next-line require-await
+    deleteKinesisEventSourceMethod: async () => {
+      throw new ResourceNotFoundException('Test Error');
+    },
+  });
+  t.deepEqual(deleteResult, [undefined, undefined]);
+});
+
 test.serial('isEventSourceMappingShared returns true if a rule shares an event source with another rule', async (t) => {
   const {
     rulePgModel,
@@ -841,7 +945,7 @@ test.serial('deleteRuleResources does not delete event source mappings if they e
   });
 });
 
-test.serial('deleteOldEventSourceMappings() removes SNS source mappings and permissions', async (t) => {
+test.serial('deleteRuleResources() removes SNS source mappings and permissions', async (t) => {
   const {
     rulePgModel,
     testKnex,
@@ -874,7 +978,7 @@ test.serial('deleteOldEventSourceMappings() removes SNS source mappings and perm
   const { Statement } = JSON.parse(Policy);
   t.true(Statement.some((s) => s.Sid === getSnsTriggerPermissionId(ruleWithTrigger)));
 
-  await deleteOldEventSourceMappings(testKnex, ruleWithTrigger);
+  await deleteRuleResources(testKnex, ruleWithTrigger);
 
   const { subExists: subExists2 } = await checkForSnsSubscriptions(ruleWithTrigger);
   t.false(subExists2);
@@ -888,7 +992,36 @@ test.serial('deleteOldEventSourceMappings() removes SNS source mappings and perm
   t.teardown(() => rulePgModel.delete(testKnex, newPgRule));
 });
 
-test.serial('deleteOldEventSourceMappings() removes kinesis source mappings', async (t) => {
+test.serial('deleteRuleResources() does not throw if a rule is passed in without a valid SNS subscription', async (t) => {
+  const {
+    testKnex,
+  } = t.context;
+
+  const topic1 = await awsServices.sns().createTopic({ Name: randomId('topic1_') }).promise();
+
+  // create rule trigger and rule
+  const snsRule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sns',
+      value: topic1.TopicArn,
+    },
+    provider: null,
+    collection: null,
+    state: 'ENABLED',
+  });
+
+  const ruleWithTrigger = await createRuleTrigger(snsRule);
+  const origSnsCheck = await checkForSnsSubscriptions(ruleWithTrigger);
+  t.true(origSnsCheck.subExists);
+
+  await awsServices.sns().unsubscribe({ SubscriptionArn: ruleWithTrigger.rule.arn }).promise();
+  const snsCheck = await checkForSnsSubscriptions(ruleWithTrigger);
+  t.false(snsCheck.subExists);
+  await t.notThrowsAsync(deleteRuleResources(testKnex, ruleWithTrigger));
+});
+
+test.serial('deleteRuleResources() removes kinesis source mappings', async (t) => {
   const {
     rulePgModel,
     testKnex,
@@ -925,7 +1058,7 @@ test.serial('deleteOldEventSourceMappings() removes kinesis source mappings', as
   t.is(consumerEventMappingsBefore.EventSourceMappings.length, 1);
   t.is(logEventMappingsBefore.EventSourceMappings.length, 1);
 
-  await deleteOldEventSourceMappings(testKnex, ruleWithTrigger);
+  await deleteRuleResources(testKnex, ruleWithTrigger);
 
   const [
     consumerEventMappingsAfter,
@@ -965,4 +1098,1061 @@ test.serial('checkForSnsSubscriptions returns the correct status of a Rule\'s su
   t.true(response.existingSubscriptionArn.includes(topic1.TopicArn));
 
   t.teardown(() => rulePgModel.delete(testKnex, newPgRule));
+});
+
+test.serial('disabling an SNS rule removes the event source mapping', async (t) => {
+  const {
+    snsTopicArn,
+    testKnex,
+  } = t.context;
+
+  const item = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sns',
+      value: snsTopicArn,
+    },
+    state: 'ENABLED',
+  });
+
+  const ruleWithTrigger = await rulesHelpers.createRuleTrigger(item);
+
+  t.is(ruleWithTrigger.rule.value, snsTopicArn);
+  t.truthy(ruleWithTrigger.rule.arn);
+  t.is(ruleWithTrigger.state, 'ENABLED');
+
+  const updates = { ...ruleWithTrigger, state: 'DISABLED' };
+  const ruleWithUpdatedTrigger = await rulesHelpers.updateRuleTrigger(
+    ruleWithTrigger,
+    updates,
+    testKnex
+  );
+
+  t.is(ruleWithUpdatedTrigger.name, ruleWithTrigger.name);
+  t.is(ruleWithUpdatedTrigger.state, 'DISABLED');
+  t.is(ruleWithUpdatedTrigger.rule.type, ruleWithTrigger.rule.type);
+  t.is(ruleWithUpdatedTrigger.rule.value, ruleWithTrigger.rule.value);
+  t.falsy(ruleWithUpdatedTrigger.rule.arn);
+});
+
+test.serial('deleting an SNS rule updates the event source mapping', async (t) => {
+  const {
+    snsTopicArn,
+    testKnex,
+  } = t.context;
+
+  const unsubscribeSpy = sinon.spy(awsServices.sns(), 'unsubscribe');
+
+  const item = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sns',
+      value: snsTopicArn,
+    },
+    state: 'ENABLED',
+  });
+
+  const ruleWithTrigger = await rulesHelpers.createRuleTrigger(item);
+
+  t.is(ruleWithTrigger.rule.value, snsTopicArn);
+
+  await rulesHelpers.deleteRuleResources(testKnex, ruleWithTrigger);
+
+  t.true(unsubscribeSpy.called);
+  t.true(unsubscribeSpy.calledWith({
+    SubscriptionArn: ruleWithTrigger.rule.arn,
+  }));
+
+  t.teardown(() => {
+    unsubscribeSpy.restore();
+  });
+});
+
+test.serial('Multiple rules using same SNS topic can be created and deleted', async (t) => {
+  const {
+    collectionPgModel,
+    providerPgModel,
+    testKnex,
+    rulePgModel,
+  } = t.context;
+  const testPgProvider = fakeProviderRecordFactory();
+  await providerPgModel.create(
+    testKnex,
+    testPgProvider,
+    '*'
+  );
+
+  const testPgCollection1 = fakeCollectionRecordFactory({
+    name: randomId('collection-'),
+    version: 'v1',
+  });
+  const testPgCollection2 = fakeCollectionRecordFactory({
+    name: randomId('collection-'),
+    version: 'v2',
+  });
+  await Promise.all([
+    collectionPgModel.create(
+      testKnex,
+      testPgCollection1,
+      '*'
+    ),
+    collectionPgModel.create(
+      testKnex,
+      testPgCollection2,
+      '*'
+    ),
+  ]);
+  const unsubscribeSpy = sinon.spy(awsServices.sns(), 'unsubscribe');
+  const { TopicArn } = await awsServices.sns().createTopic({
+    Name: randomId('topic'),
+  }).promise();
+
+  const ruleWithTrigger = await rulesHelpers.createRuleTrigger(fakeRuleFactoryV2({
+    name: randomId('rule1'),
+    rule: {
+      type: 'sns',
+      value: TopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+    collection: {
+      name: testPgCollection1.name,
+      version: testPgCollection1.version,
+    },
+    provider: testPgProvider.name,
+  }));
+  const ruleWithTrigger2 = await rulesHelpers.createRuleTrigger(fakeRuleFactoryV2({
+    name: randomId('rule2'),
+    rule: {
+      type: 'sns',
+      value: TopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+    collection: {
+      name: testPgCollection2.name,
+      version: testPgCollection2.version,
+    },
+    provider: testPgProvider.name,
+  }));
+
+  const pgRule = await translateApiRuleToPostgresRule(ruleWithTrigger, testKnex);
+  const pgRule2 = await translateApiRuleToPostgresRule(ruleWithTrigger2, testKnex);
+
+  const [rule1] = await rulePgModel.create(testKnex, pgRule);
+  const [rule2] = await rulePgModel.create(testKnex, pgRule2);
+
+  // rules share the same subscription
+  t.is(rule1.arn, rule2.arn);
+
+  // Have to delete rules serially otherwise all rules still exist
+  // when logic to check for shared source mapping is evaluated
+
+  await rulesHelpers.deleteSnsTrigger(testKnex, ruleWithTrigger);
+  await rulePgModel.delete(testKnex, rule1);
+
+  await t.notThrowsAsync(rulesHelpers.deleteSnsTrigger(testKnex, ruleWithTrigger2));
+  await rulePgModel.delete(testKnex, rule2);
+
+  // Ensure that cleanup for SNS rule subscription was actually called
+  t.true(unsubscribeSpy.called);
+  t.true(unsubscribeSpy.calledWith({
+    SubscriptionArn: ruleWithTrigger.rule.arn,
+  }));
+
+  t.teardown(async () => {
+    unsubscribeSpy.restore();
+    await awsServices.sns().deleteTopic({
+      TopicArn,
+    });
+  });
+});
+
+test.serial('deleteSnsTrigger throws more detailed ResourceNotFoundError', async (t) => {
+  const errorMessage = 'Resource is not found in resource policy.';
+  const error = new Error(errorMessage);
+  error.code = 'ResourceNotFoundException';
+  const { snsTopicArn } = t.context;
+  const lambdaStub = sinon.stub(awsServices.lambda(), 'removePermission').throws(error);
+
+  const ruleWithTrigger = await rulesHelpers.createRuleTrigger(fakeRuleFactoryV2({
+    rule: {
+      type: 'sns',
+      value: snsTopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+  }));
+
+  await t.throwsAsync(
+    rulesHelpers.deleteSnsTrigger(t.context.testKnex, ruleWithTrigger), {
+      instanceOf: ResourceNotFoundError,
+      message: `${errorMessage} ${resourceNotFoundInfo}`,
+    }
+  );
+
+  t.teardown(() => {
+    lambdaStub.restore();
+  });
+});
+
+// ***Create Rules Resource Tests
+test('Creating a rule trigger defaults enabled to true', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'onetime',
+    },
+    workflow,
+  });
+  // remove enabled from rule to be created
+  delete rule.state;
+
+  // create rule trigger
+  const ruleWithTrigger = await createRuleTrigger(rule);
+
+  t.is(ruleWithTrigger.state, 'ENABLED');
+});
+
+test('Creating an invalid rule does not create workflow triggers', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'onetime',
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  rule.rule.type = 'invalid';
+
+  await t.throwsAsync(
+    () => createRuleTrigger(rule),
+    { name: 'ValidationError' }
+  );
+});
+
+test('Creating a rule trigger for a onetime rule succeeds', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'onetime',
+    },
+    state: 'ENABLED',
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+
+  const onetimeRule = await createRuleTrigger(rule);
+
+  t.deepEqual(onetimeRule, rule);
+});
+
+test.serial('Creating rule triggers for a kinesis type rule adds event mappings', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+  });
+  // create rule
+  const createdRule = await createRuleTrigger(kinesisRule);
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 1);
+  t.is(logEventMappings.length, 1);
+  t.is(consumerEventMappings[0].UUID, createdRule.rule.arn);
+  t.is(logEventMappings[0].UUID, createdRule.rule.logEventArn);
+
+  t.is(createdRule.name, kinesisRule.name);
+  t.is(createdRule.rule.value, kinesisRule.rule.value);
+  t.false(createdRule.rule.arn === undefined);
+  t.false(createdRule.rule.logEventArn === undefined);
+});
+
+test.serial('Repeatedly creating rule triggers for a kinesis type rule utilizes the same event mappings', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+  });
+  // create rule
+  const originalCreatedRule = await createRuleTrigger(kinesisRule);
+  await createRuleTrigger(kinesisRule);
+
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 1);
+  t.is(logEventMappings.length, 1);
+  t.is(consumerEventMappings[0].UUID, originalCreatedRule.rule.arn);
+  t.is(logEventMappings[0].UUID, originalCreatedRule.rule.logEventArn);
+
+  t.is(originalCreatedRule.name, kinesisRule.name);
+  t.is(originalCreatedRule.rule.value, kinesisRule.rule.value);
+  t.false(originalCreatedRule.rule.arn === undefined);
+  t.false(originalCreatedRule.rule.logEventArn === undefined);
+});
+
+test.serial('Recreating rule triggers for a kinesis type rule over a disabled rule utilizes the same event mappings', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+  });
+
+  // Create rule triggers
+  const originalCreatedRule = await createRuleTrigger(kinesisRule);
+  const originalKinesisEventMappings = await getKinesisEventMappings();
+  // Set event mapping to disabled
+  originalKinesisEventMappings.map((mapping) =>
+    awsServices
+      .lambda()
+      .updateEventSourceMapping({
+        UUID: mapping.EventSourceMappings[0].UUID,
+        Enabled: false,
+      })
+      .promise());
+
+  // Recreate rule triggers
+  await createRuleTrigger(kinesisRule);
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 1);
+  t.is(logEventMappings.length, 1);
+  t.is(consumerEventMappings[0].UUID, originalCreatedRule.rule.arn);
+  t.is(logEventMappings[0].UUID, originalCreatedRule.rule.logEventArn);
+
+  t.is(originalCreatedRule.name, kinesisRule.name);
+  t.is(originalCreatedRule.rule.value, kinesisRule.rule.value);
+  t.false(originalCreatedRule.rule.arn === undefined);
+  t.false(originalCreatedRule.rule.logEventArn === undefined);
+});
+
+test.serial('Creating an invalid kinesis type rule does not add event mappings', async (t) => {
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: randomString(),
+    },
+  });
+  delete kinesisRule.name;
+  await t.throwsAsync(createRuleTrigger(kinesisRule), { name: 'SchemaValidationError' });
+
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 0);
+  t.is(logEventMappings.length, 0);
+});
+
+test('Creating a rule trigger SQS rule fails if queue does not exist', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'sqs',
+      value: 'non-existent-queue',
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  await t.throwsAsync(
+    createRuleTrigger(rule),
+    { message: /SQS queue non-existent-queue does not exist/ }
+  );
+});
+
+test('Creating a rule trigger for an SQS rule fails if there is no redrive policy on the queue', async (t) => {
+  const queueUrl = await SQS.createQueue(randomId('queue'));
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sqs',
+      value: queueUrl,
+    },
+    state: 'ENABLED',
+  });
+  await t.throwsAsync(
+    createRuleTrigger(rule),
+    { message: `SQS queue ${queueUrl} does not have a dead-letter queue configured` }
+  );
+  t.teardown(async () => await SQS.deleteQueue(queueUrl));
+});
+
+test('Creating a rule trigger for an SQS rule succeeds', async (t) => {
+  const queues = await createSqsQueues(randomString());
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sqs',
+      value: queues.queueUrl,
+    },
+    state: 'ENABLED',
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+  const sqsRule = await createRuleTrigger(rule);
+  t.deepEqual(sqsRule, rule);
+  t.teardown(async () => await SQS.deleteQueue(queues.queueUrl));
+});
+
+test('Creating a rule trigger for an SQS rule succeeds and sets default value for meta.retries and visibilityTimeout', async (t) => {
+  const queues = await createSqsQueues(randomString());
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sqs',
+      value: queues.queueUrl,
+    },
+    state: 'ENABLED',
+  });
+  const sqsRule = await createRuleTrigger(rule);
+  t.is(sqsRule.meta.retries, 3);
+  t.is(sqsRule.meta.visibilityTimeout, 300);
+  t.teardown(async () => await SQS.deleteQueue(queues.queueUrl));
+});
+
+test('Creating a rule trigger for a rule without a type fails', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'onetime',
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  delete rule.rule.type;
+
+  await t.throwsAsync(
+    () => createRuleTrigger(rule),
+    { name: 'SchemaValidationError' }
+  );
+});
+
+test('Creating a rule trigger for a rule without a workflow fails', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'onetime',
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  delete rule.workflow;
+
+  await t.throwsAsync(
+    () => createRuleTrigger(rule),
+    { name: 'SchemaValidationError' }
+  );
+});
+
+test('Creating a rule trigger for a rule without a name fails', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'onetime',
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  delete rule.name;
+
+  await t.throwsAsync(
+    () => createRuleTrigger(rule),
+    { name: 'SchemaValidationError' }
+  );
+});
+
+test('Creating a disabled SNS rule creates no event source mapping', async (t) => {
+  const snsTopicArn = randomString();
+  const item = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sns',
+      value: snsTopicArn,
+    },
+    state: 'DISABLED',
+  });
+
+  const rule = await createRuleTrigger(item);
+
+  t.is(rule.state, 'DISABLED');
+  t.is(rule.rule.value, snsTopicArn);
+  t.falsy(rule.rule.arn);
+});
+
+test.serial('Creating an enabled SNS rule creates an event source mapping', async (t) => {
+  const { TopicArn } = await awsServices.sns().createTopic({
+    Name: randomId('topic'),
+  }).promise();
+
+  const lambdaStub = sinon.stub(awsServices, 'lambda')
+    .returns({
+      addPermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+      removePermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+
+  const snsStub = sinon.stub(awsServices, 'sns')
+    .returns({
+      listSubscriptionsByTopic: () => ({
+        promise: () => Promise.resolve({
+          Subscriptions: [{
+            SubscriptionArn: randomString(),
+          }],
+        }),
+      }),
+      subscribe: () => ({
+        promise: () => Promise.resolve({
+          SubscriptionArn: randomString(),
+        }),
+      }),
+    });
+
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'sns',
+      value: TopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+  const subscribeSpy = sinon.spy(awsServices.sns(), 'subscribe');
+  const addPermissionSpy = sinon.spy(awsServices.lambda(), 'addPermission');
+
+  await createRuleTrigger(rule);
+  t.true(subscribeSpy.called);
+  t.true(subscribeSpy.calledWith({
+    TopicArn: rule.rule.value,
+    Protocol: 'lambda',
+    Endpoint: process.env.messageConsumer,
+    ReturnSubscriptionArn: true,
+  }));
+  t.true(addPermissionSpy.called);
+  t.teardown(async () => {
+    lambdaStub.restore();
+    snsStub.restore();
+    subscribeSpy.restore();
+    addPermissionSpy.restore();
+    await awsServices.sns().deleteTopic({ TopicArn }).promise();
+  });
+});
+
+test.serial('Creating a rule trigger for a scheduled rule succeeds', async (t) => {
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'scheduled',
+      value: 'rate(1 min)',
+    },
+    state: 'ENABLED',
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+
+  const cloudwatchStub = sinon.stub(awsServices, 'cloudwatchevents')
+    .returns({
+      putRule: () => ({
+        promise: () => Promise.resolve(),
+      }),
+      putTargets: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+  const scheduledRule = await createRuleTrigger(rule);
+  t.true(cloudwatchStub.called);
+  t.deepEqual(scheduledRule, rule);
+  t.teardown(() => {
+    cloudwatchStub.restore();
+  });
+});
+
+test('buildPayload builds a lambda payload from the rule', async (t) => {
+  const collectionPgModel = new CollectionPgModel();
+  const providerPgModel = new ProviderPgModel();
+  const testPgProvider = fakeProviderRecordFactory();
+  const [pgProvider] = await providerPgModel.create(
+    t.context.testKnex,
+    testPgProvider,
+    '*'
+  );
+  const collectionName = 'fakeCollection';
+  const collectionVersion = 'v1';
+  const testPgCollection = fakeCollectionRecordFactory({
+    name: collectionName,
+    version: collectionVersion,
+  });
+  const [pgCollection] = await collectionPgModel.create(
+    t.context.testKnex,
+    testPgCollection
+  );
+  const ruleParams = {
+    rule: { type: 'onetime' },
+    workflow,
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+    cumulus_meta: {
+      execution_name: 'fakeName',
+    },
+    payload: {
+      input: 'test',
+    },
+    queueUrl: randomString(),
+    executionNamePrefix: randomString(),
+    asyncOperationId: randomString(),
+    provider: pgProvider.name,
+    collection: { name: pgCollection.name, version: pgCollection.version },
+  };
+  const rule = fakeRuleFactoryV2(ruleParams);
+  const expectedPayload = {
+    provider: pgProvider.name,
+    collection: { name: pgCollection.name, version: pgCollection.version },
+    meta: rule.meta,
+    cumulus_meta: { execution_name: rule.cumulus_meta.execution_name },
+    payload: rule.payload,
+    queueUrl: rule.queueUrl,
+    executionNamePrefix: rule.executionNamePrefix,
+    asyncOperationId: rule.asyncOperationId,
+  };
+  const payload = await buildPayload(rule, ruleParams.cumulus_meta, ruleParams.asyncOperationId);
+  t.deepEqual(omit(payload, ['template', 'definition']), expectedPayload);
+});
+
+test('buildPayload throws error if workflow file does not exist', async (t) => {
+  const fakeWorkflow = randomString();
+  const workflowFileKey = workflows.getWorkflowFileKey(process.env.stackName, fakeWorkflow);
+  const ruleParams = {
+    rule: { type: 'onetime' },
+    workflow: fakeWorkflow,
+  };
+  const rule = fakeRuleFactoryV2(ruleParams);
+  await t.throwsAsync(
+    buildPayload(rule),
+    { message: `Workflow doesn\'t exist: s3://${process.env.system_bucket}/${workflowFileKey} for ${rule.name}` }
+  );
+});
+
+test.serial('Updating a rule trigger with an "onetime" rule type returns updated rule', async (t) => {
+  const fakeRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'onetime',
+      field: 'testField',
+    },
+  });
+  // create rule trigger
+  const createdRule = await createRuleTrigger(fakeRule);
+  const updatedRule = { ...createdRule, field: 'newTestField' };
+  const updatedRuleTriggerOutput = await updateRuleTrigger(createdRule, updatedRule);
+  t.deepEqual(updatedRuleTriggerOutput, updatedRule);
+});
+
+test.serial('Updating a rule trigger with an invalid rule type throws an error', async (t) => {
+  const fakeRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'totallyNotARule',
+    },
+  });
+  // create rule trigger
+  await t.throwsAsync(createRuleTrigger(fakeRule), { name: 'ValidationError' });
+});
+
+// update rule resources tests
+test.serial('Updating a kinesis type rule workflow does not affect value or event source mappings', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const { testKnex } = t.context;
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+  });
+  // create rule trigger
+  const createdRule = await createRuleTrigger(kinesisRule);
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 1);
+  t.is(logEventMappings.length, 1);
+  t.is(consumerEventMappings[0].UUID, createdRule.rule.arn);
+  t.is(logEventMappings[0].UUID, createdRule.rule.logEventArn);
+
+  // update rule workflow
+  const updatedKinesisRule = {
+    ...createdRule,
+    workflow: 'new-workflow',
+  };
+  const updatedRule = await updateRuleTrigger(kinesisRule, updatedKinesisRule, testKnex);
+  const updatedKinesisEventMappings = await getKinesisEventMappings();
+  const updatedconsumerEventMappings = updatedKinesisEventMappings[0].EventSourceMappings;
+  const updatedlogEventMappings = updatedKinesisEventMappings[1].EventSourceMappings;
+
+  t.is(updatedconsumerEventMappings.length, 1);
+  t.is(updatedlogEventMappings.length, 1);
+  t.is(updatedconsumerEventMappings[0].UUID, updatedRule.rule.arn);
+  t.is(updatedlogEventMappings[0].UUID, updatedRule.rule.logEventArn);
+
+  t.is(createdRule.rule.arn, updatedRule.rule.arn);
+  t.is(createdRule.rule.logEventArn, updatedRule.rule.logEventArn);
+
+  // Clean Up
+  t.teardown(async () => {
+    await deleteRuleResources(testKnex, createdRule);
+    await deleteRuleResources(testKnex, updatedRule);
+  });
+});
+
+test.serial('Updating a kinesis type rule workflow with new arn updates event source mappings', async (t) => {
+  const kinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const updatedKinesisArn = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const { testKnex } = t.context;
+  const kinesisRule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn,
+    },
+  });
+  // create rule trigger
+  const createdRule = await createRuleTrigger(kinesisRule);
+  const kinesisEventMappings = await getKinesisEventMappings();
+  const consumerEventMappings = kinesisEventMappings[0].EventSourceMappings;
+  const logEventMappings = kinesisEventMappings[1].EventSourceMappings;
+
+  t.is(consumerEventMappings.length, 1);
+  t.is(logEventMappings.length, 1);
+  t.is(consumerEventMappings[0].UUID, createdRule.rule.arn);
+  t.is(logEventMappings[0].UUID, createdRule.rule.logEventArn);
+
+  // update rule workflow
+  const updatedKinesisRule = {
+    ...createdRule,
+    workflow: 'new-workflow',
+    rule: { ...createdRule.rule, value: updatedKinesisArn },
+  };
+  const updatedRule = await updateRuleTrigger(kinesisRule, updatedKinesisRule, testKnex);
+  const updatedKinesisEventMappings = await getKinesisEventMappings();
+  const updatedconsumerEventMappings = updatedKinesisEventMappings[0].EventSourceMappings;
+  const updatedlogEventMappings = updatedKinesisEventMappings[1].EventSourceMappings;
+
+  t.is(updatedconsumerEventMappings.length, 2);
+  t.is(updatedlogEventMappings.length, 2);
+
+  t.is(updatedconsumerEventMappings[1].UUID, updatedRule.rule.arn);
+  t.is(updatedlogEventMappings[1].UUID, updatedRule.rule.logEventArn);
+  t.is(updatedconsumerEventMappings[0].UUID, createdRule.rule.arn);
+  t.is(updatedlogEventMappings[0].UUID, createdRule.rule.logEventArn);
+
+  // Clean Up
+  t.teardown(async () => {
+    await deleteRuleResources(testKnex, createdRule);
+    await deleteRuleResources(testKnex, updatedRule);
+  });
+});
+
+test.serial('Updating a valid SQS rule to have an invalid schema throws an error', async (t) => {
+  const queues = await createSqsQueues(randomString());
+  const sqsRule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sqs',
+      value: queues.queueUrl,
+    },
+    state: 'ENABLED',
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+  const sqsRuleWithTrigger = await createRuleTrigger(sqsRule);
+  t.is(sqsRuleWithTrigger.value, sqsRule.value);
+
+  // update rule to be invalid by setting type to null
+  const updatedSqsRule = {
+    ...sqsRuleWithTrigger,
+    rule: {
+      ...sqsRuleWithTrigger.rule,
+      type: null,
+    },
+  };
+  await t.throwsAsync(
+    updateRuleTrigger(sqsRuleWithTrigger, updatedSqsRule, t.context.testKnex),
+    { name: 'SchemaValidationError' }
+  );
+  t.teardown(async () => {
+    await deleteRuleResources(t.context.testKnex, sqsRuleWithTrigger);
+    await SQS.deleteQueue(queues.queueUrl);
+  });
+});
+
+test.serial('Updating an SQS rule fails if there is no redrive policy on the queue', async (t) => {
+  const queues = await createSqsQueues(randomString());
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    rule: {
+      type: 'sqs',
+      value: queues.queueUrl,
+    },
+    state: 'ENABLED',
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+  const sqsRule = await createRuleTrigger(rule);
+  const queueUrl = await SQS.createQueue(randomId('queue'));
+  const updatedRule = {
+    ...rule,
+    rule: {
+      ...rule.rule,
+      value: queueUrl,
+    },
+  };
+  await t.throwsAsync(
+    updateRuleTrigger(sqsRule, updatedRule, t.context.testKnex),
+    { message: `SQS queue ${queueUrl} does not have a dead-letter queue configured` }
+  );
+  t.teardown(async () => await SQS.deleteQueue(queueUrl));
+});
+
+test.serial('Updating the queue for an SQS rule succeeds', async (t) => {
+  const queues = await createSqsQueues(randomString());
+  const rule = fakeRuleFactoryV2({
+    workflow,
+    state: 'ENABLED',
+    rule: {
+      type: 'sqs',
+      value: queues.queueUrl,
+    },
+    meta: {
+      visibilityTimeout: 100,
+      retries: 4,
+    },
+  });
+  const sqsRule = await createRuleTrigger(rule);
+  t.deepEqual(sqsRule.rule.value, queues.queueUrl);
+
+  const newQueues = await createSqsQueues(randomString());
+  const updatedRule = {
+    ...sqsRule,
+    rule: {
+      ...rule.rule,
+      value: newQueues.queueUrl,
+    },
+  };
+  const updatedSqsRule = await updateRuleTrigger(sqsRule, updatedRule, t.context.testKnex);
+  t.is(updatedSqsRule.rule.value, newQueues.queueUrl);
+  t.teardown(async () => {
+    await SQS.deleteQueue(queues.queueUrl);
+    await SQS.deleteQueue(newQueues.queueUrl);
+  });
+});
+
+test.serial('Updating an SNS rule updates the event source mapping', async (t) => {
+  const snsTopicArn = randomString();
+  const newSnsTopicArn = randomString();
+  const { TopicArn } = await awsServices.sns().createTopic({
+    Name: snsTopicArn,
+  }).promise();
+  const { TopicArn: TopicArn2 } = await awsServices.sns().createTopic({
+    Name: newSnsTopicArn,
+  }).promise();
+
+  const lambdaStub = sinon.stub(awsServices, 'lambda')
+    .returns({
+      addPermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+      removePermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+  const snsStub = sinon.stub(awsServices, 'sns')
+    .returns({
+      listSubscriptionsByTopic: () => ({
+        promise: () => Promise.resolve({
+          Subscriptions: [{
+            Endpoint: process.env.messageConsumer,
+            SubscriptionArn: randomString(),
+          }],
+        }),
+      }),
+      unsubscribe: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'sns',
+      value: TopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+
+  const ruleWithTrigger = await createRuleTrigger(rule);
+
+  t.is(ruleWithTrigger.rule.value, TopicArn);
+  t.truthy(ruleWithTrigger.rule.arn);
+
+  const updates = {
+    ...rule,
+    rule: {
+      ...rule.rule,
+      arn: ruleWithTrigger.rule.arn,
+      value: TopicArn2,
+    },
+  };
+  const updatedSnsRule = await updateRuleTrigger(ruleWithTrigger, updates, t.context.testKnex);
+
+  t.is(updatedSnsRule.name, rule.name);
+  t.is(updatedSnsRule.rule.type, rule.rule.type);
+  t.is(updatedSnsRule.rule.value, TopicArn2);
+  t.not(updatedSnsRule.rule.arn, rule.rule.arn);
+
+  t.teardown(async () => {
+    lambdaStub.restore();
+    snsStub.restore();
+    await awsServices.sns().deleteTopic({ TopicArn: TopicArn2 }).promise();
+  });
+});
+
+test.serial('Updating an SNS rule to "disabled" removes the event source mapping ARN', async (t) => {
+  const snsTopicArn = randomString();
+  const { TopicArn } = await awsServices.sns().createTopic({
+    Name: snsTopicArn,
+  }).promise();
+
+  const lambdaStub = sinon.stub(awsServices, 'lambda')
+    .returns({
+      addPermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+      removePermission: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+  const snsStub = sinon.stub(awsServices, 'sns')
+    .returns({
+      listSubscriptionsByTopic: () => ({
+        promise: () => Promise.resolve({
+          Subscriptions: [{
+            Endpoint: process.env.messageConsumer,
+            SubscriptionArn: randomString(),
+          }],
+        }),
+      }),
+      unsubscribe: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'sns',
+      value: TopicArn,
+    },
+    workflow,
+    state: 'ENABLED',
+  });
+
+  const ruleWithTrigger = await createRuleTrigger(rule);
+
+  t.is(ruleWithTrigger.rule.value, TopicArn);
+  t.truthy(ruleWithTrigger.rule.arn);
+
+  const updates = {
+    ...ruleWithTrigger,
+    state: 'DISABLED',
+  };
+  const updatedSnsRule = await updateRuleTrigger(ruleWithTrigger, updates, t.context.testKnex);
+
+  t.true(Object.prototype.hasOwnProperty.call(updatedSnsRule.rule, 'arn'));
+  t.is(updatedSnsRule.rule.arn, undefined);
+
+  t.teardown(async () => {
+    lambdaStub.restore();
+    snsStub.restore();
+    await awsServices.sns().deleteTopic({ TopicArn }).promise();
+  });
+});
+
+test.serial('Enabling a disabled SNS rule and passing rule.arn throws specific error', async (t) => {
+  const snsTopicArn = randomString();
+  const snsStub = sinon.stub(awsServices, 'sns')
+    .returns({
+      listSubscriptionsByTopic: () => ({
+        promise: () => Promise.resolve({
+          Subscriptions: [{
+            Endpoint: process.env.messageConsumer,
+            SubscriptionArn: randomString(),
+          }],
+        }),
+      }),
+      unsubscribe: () => ({
+        promise: () => Promise.resolve(),
+      }),
+    });
+
+  const rule = fakeRuleFactoryV2({
+    rule: {
+      type: 'sns',
+      value: snsTopicArn,
+    },
+    workflow,
+    state: 'DISABLED',
+  });
+
+  t.is(rule.rule.value, snsTopicArn);
+  t.falsy(rule.rule.arn);
+  t.is(rule.state, 'DISABLED');
+
+  const updates = {
+    name: rule.name,
+    state: 'ENABLED',
+    rule: {
+      ...rule.rule,
+      arn: 'test-value',
+    },
+  };
+
+  // Should fail because a disabled rule should not have an ARN
+  // when being updated
+  await t.throwsAsync(updateRuleTrigger(rule, updates, t.context.testKnex),
+    null,
+    'Including rule.arn is not allowed when enabling a disabled rule');
+  t.teardown(() => {
+    snsStub.restore();
+  });
 });
